removed warning
git-svn-id: http://svn.seqan.de/seqan/trunk@6361 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/library/seqan/index/find_swift.h
@@ -887,7 +887,7 @@ inline bool _swiftMultiProcessQGram(
 				hitCount = 1;
 				(*bkt).firstIncrement = curPos;
 			}
-			else if((*bkt).lastIncrement + bucketParams.tabooLength > curPos)
+			else if((__int64)((*bkt).lastIncrement + bucketParams.tabooLength) > curPos)
 			{
 				// bkt counter was already incremented for another q-gram at
 				//   a haystack position that is closer than tabooLength
